ViewLayer Class implemented
